gen4 planner refactoring
