+++ go/vt/vtgate/planbuilder/operators/operator.go
@@ -83,80 +83,6 @@ func PlanQuery(ctx *plancontext.PlanningContext, selStmt sqlparser.Statement) (o
 	return op, err
 }
 
-func tryHorizonPlanning(ctx *plancontext.PlanningContext, root ops.Operator) (output ops.Operator, err error) {
-	backup := Clone(root)
-	defer func() {
-		// If we get the below error.
-		// We will be using old horizon planning.
-		if err == _errHorizonNotPlanned {
-			err = planOffsetsOnJoins(ctx, backup)
-			if err == nil {
-				output = backup
-			}
-		}
-	}()
-
-	_, ok := root.(*Horizon)
-
-	if !ok || len(ctx.SemTable.SubqueryMap) > 0 || len(ctx.SemTable.SubqueryRef) > 0 {
-		// we are not ready to deal with subqueries yet
-		return root, errHorizonNotPlanned()
-	}
-
-	output, err = planHorizons(ctx, root)
-	if err != nil {
-		return nil, err
-	}
-
-	output, err = planOffsets(ctx, output)
-	if err != nil {
-		return nil, err
-	}
-
-	output, err = makeSureOutputIsCorrect(ctx, root, output)
-	if err != nil {
-		return nil, err
-	}
-
-	return
-}
-
-func makeSureOutputIsCorrect(ctx *plancontext.PlanningContext, oldHorizon ops.Operator, output ops.Operator) (ops.Operator, error) {
-	// next we use the original Horizon to make sure that the output columns line up with what the user asked for
-	// in the future, we'll tidy up the results. for now, we are just failing these queries and going back to the
-	// old horizon planning instead
-	cols, err := output.GetColumns()
-	if err != nil {
-		return nil, err
-	}
-
-	horizon := oldHorizon.(*Horizon)
-
-	sel := sqlparser.GetFirstSelect(horizon.Select)
-
-	if len(sel.SelectExprs) == len(cols) {
-		return output, nil
-	}
-
-	if tryTruncateColumnsAt(output, len(sel.SelectExprs)) {
-		return output, nil
-	}
-
-	qp, err := horizon.getQP(ctx)
-	if err != nil {
-		return nil, err
-	}
-	proj, err := createProjection2(qp, output)
-	if err != nil {
-		return nil, err
-	}
-	err = proj.passThroughAllColumns(ctx)
-	if err != nil {
-		return nil, err
-	}
-	return proj, nil
-}
-
 // Inputs implements the Operator interface
 func (noInputs) Inputs() []ops.Operator {
 	return nil
